100-get_endianness.c

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -1,9 +1,13 @@
 #include "main.h"
-#include <stdint.h>
 
+/**
+ * get_endianness - checks if a machine is little or big endian
+ * Return: 0 for big, 1 for little
+ */
 int get_endianness(void)
 {
-	uint32_t i = 1;
+	unsigned int i = 1;
 	char *c = (char *) &i;
-	 return (int) (*c);
+
+	return (*c);
 }
